Not working properly
Due to Mixin related errors and other event hooking errors, this mod will probably not work properly in the future. I have used various tricks to get it to work properly in a development environment (still a few errors), but in a production environment it was no longer as good as it could be.